cast uint16_t to size_t for ADUC_AllocAndStrCopyN

+++ src/extensions/agent_modules/shared/mosquitto_utils/src/adu_mosquitto_utils.c
@@ -98,7 +98,7 @@ bool ADU_are_correlation_ids_matching(const mosquitto_property* props, const cha
 
     if (outCorrelationData != NULL)
     {
-        if (ADUC_AllocAndStrCopyN(outCorrelationData, cid, cidLen) != 0)
+        if (ADUC_AllocAndStrCopyN(outCorrelationData, cid, (size_t)cidLen) != 0)
         {
             goto done;
         }
